# backend/docs/swagger.yml
openapi: 3.0.3
info:
  title: QMin Backend API
  version: 1.0.0
  description: API for task management in the QMin backend.
servers:
  - url: http://localhost:5000
paths:
  /health:
    get:
      summary: Server health check
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: OK
                  timestamp:
                    type: string
                    format: date-time
                    example: '2026-05-17T15:49:02.055Z'
  /api/tasks:
    get:
      summary: Get all tasks
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
    get:
      summary: Get a single task by ID
      responses:
        '200':
          description: Task returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete a task
      responses:
        '200':
          description: Task deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Task deleted successfully
                  task:
                    $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/tasks/status/{status}:
    parameters:
      - name: status
        in: path
        required: true
        schema:
          type: string
          enum:
            - To-Do
            - In Progress
            - Done
    get:
      summary: Get tasks by status
      responses:
        '200':
          description: Tasks filtered by status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/tasks/{id}/status:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
    patch:
      summary: Update only the task status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskStatusRequest'
      responses:
        '200':
          description: Task status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Build frontend
        description:
          type: string
          example: Create SvelteKit app
        status:
          type: string
          enum:
            - To-Do
            - In Progress
            - Done
          example: In Progress
        created_at:
          type: string
          format: date-time
          example: '2026-05-17T15:40:00.000Z'
        updated_at:
          type: string
          format: date-time
          example: '2026-05-17T15:42:00.000Z'
    CreateTaskRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Build frontend
        description:
          type: string
          example: Create SvelteKit app
    UpdateTaskRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Build frontend
        description:
          type: string
          example: Add UI for tasks
        status:
          type: string
          enum:
            - To-Do
            - In Progress
            - Done
          example: Done
    UpdateTaskStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - To-Do
            - In Progress
            - Done
          example: In Progress
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid status
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Task not found
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Failed to fetch tasks